move connect command to irc mod
