+++ src-tauri/src/irc/mod.rs
@@ -10,5 +10,43 @@ pub mod message;
 pub mod websocket;
 
 pub use client::IrcClient;
-pub use config::ClientConfig;
-pub use error::Error;
+use config::ClientConfig;
+use error::Error;
+use message::ServerMessage;
+use tauri::ipc::Channel;
+use tauri::{async_runtime, State};
+use tokio::sync::Mutex;
+
+use crate::api::get_access_token;
+use crate::error::Error as AppError;
+use crate::AppState;
+
+#[tauri::command]
+pub async fn connect(
+    state: State<'_, Mutex<AppState>>,
+    channel: Channel<ServerMessage>,
+) -> Result<(), AppError> {
+    let mut guard = state.lock().await;
+    let token = get_access_token(&guard).await?;
+    let login = token.login.to_string();
+
+    let config = ClientConfig::new(
+        login.clone(),
+        // Need to convert to &str first because AccessToken::to_string masks
+        // the actual token
+        token.access_token.as_str().to_string(),
+    );
+
+    let (mut incoming, client) = IrcClient::new(config);
+
+    async_runtime::spawn(async move {
+        while let Some(message) = incoming.recv().await {
+            channel.send(message).unwrap();
+        }
+    });
+
+    client.connect().await;
+    guard.irc = Some(client);
+
+    Ok(())
+}
